+++ chain/chain/src/stateless_validation/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod chunk_endorsement;
+mod metrics;
 pub(crate) mod state_transition_data;
 pub mod state_witness;
